fix: use safe path for tmp/temp and allow pipeline input
This includes one functional fix, and two minor details.

1) Safe method for getting path to temp directory

Using $env:TEMP is a Windows-only way of getting the temp directory.
By replacing this with [System.Io.Path]::GetTempPath() it allows
the Convert-JsonToBicep function to work across all platforms.

2) Allow Convert-JsonToBicep to receive pipeline input

This allows the function to process pipeline input to be used
in common Powershell scripting patterns.

3) Trailing whitespace removals (editor formatting, yay)